+++ src/cli/commands/analyze/dependency-risk-helpers/table/format-dependency-risks-table.ts
@@ -18,50 +18,21 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import { bold, dim, green, red, STATUS_ICONS } from '../../../../ui/colors.js';
-import { effectiveStatus, sortReleases } from './analysis-response.ts';
+import { dim, green, red, STATUS_ICONS } from '../../../../../ui/colors.js';
+import { sortReleases } from '../analysis-response.ts';
 import type {
   AnalysisErrorResource,
   AnalyzeProjectIssue,
   AnalyzeProjectRelease,
   AnalyzeProjectResponse,
   ScaIssueType,
-  VersionOption,
-  VersionOptionDescriptionCode,
-} from './sca-scanner.ts';
+} from '../sca-scanner.ts';
+import { appendLicenseGroup } from './format-table-license-group.ts';
+import { appendMalwareGroup } from './format-table-malware-group.ts';
+import { appendVulnerabilityGroup } from './format-table-vulnerability-group.ts';
 
-const DESCRIPTION_CODE_ORDER: Record<VersionOptionDescriptionCode, number> = {
-  LATEST_STABLE: 0,
-  LATEST_COMPLETE: 1,
-  LATEST_PRERELEASE: 2,
-  LATEST_PARTIAL: 3,
-  NEAREST_COMPLETE: 4,
-  NEAREST_PARTIAL: 5,
-  VERSION_IN_USE: 6,
-  UNKNOWN: 7,
-};
-
-const EXCLUDED_DESCRIPTION_CODES: ReadonlySet<VersionOptionDescriptionCode> = new Set([
-  'VERSION_IN_USE',
-  'UNKNOWN',
-]);
-
-const DESCRIPTION_CODE_LABEL: Record<VersionOptionDescriptionCode, string | null> = {
-  LATEST_STABLE: 'latest stable',
-  LATEST_COMPLETE: 'latest',
-  LATEST_PRERELEASE: 'latest prerelease',
-  LATEST_PARTIAL: 'latest',
-  NEAREST_COMPLETE: 'nearest',
-  NEAREST_PARTIAL: 'nearest',
-  VERSION_IN_USE: null,
-  UNKNOWN: null,
-};
-
-const SEVERITY_WIDTH = 9;
-const STATUS_WIDTH = 8;
 const MAX_LINE_WIDTH = 80;
 const MAX_CHAINS_DISPLAYED = 3;
-const MAX_PACKAGE_FIXES = 2;
 const CHAIN_LINE_INDENT = '  ';
 const CHAIN_CONTINUATION_INDENT = `${CHAIN_LINE_INDENT}    `;
 
@@ -167,34 +138,6 @@ function emitGroup(
   }
 }
 
-function appendMalwareGroup(
-  lines: string[],
-  release: AnalyzeProjectRelease,
-  malware: AnalyzeProjectIssue[],
-): void {
-  for (const issue of malware) lines.push(issueLine(release, issue));
-  lines.push(red(bold('Remove this package and notify your information security team')));
-}
-
-function appendLicenseGroup(
-  lines: string[],
-  release: AnalyzeProjectRelease,
-  licenses: AnalyzeProjectIssue[],
-): void {
-  for (const issue of licenses) lines.push(issueLine(release, issue));
-  lines.push(bold('Review the license usage'));
-}
-
-function appendVulnerabilityGroup(
-  lines: string[],
-  release: AnalyzeProjectRelease,
-  vulnerabilities: AnalyzeProjectIssue[],
-): void {
-  for (const issue of vulnerabilities) lines.push(issueLine(release, issue));
-  const fixLine = packageFixLine(vulnerabilities);
-  if (fixLine !== null) lines.push(fixLine);
-}
-
 function getLabel(release: AnalyzeProjectRelease): string {
   return `${release.packageName}@${release.version}`;
 }
@@ -208,122 +151,6 @@ function packageHeader(release: AnalyzeProjectRelease): string {
     return `${label}─`;
   }
   return label + '─'.repeat(MAX_LINE_WIDTH - label.length);
-}
-
-function issueLine(release: AnalyzeProjectRelease, issue: AnalyzeProjectIssue): string {
-  const severity = issue.severity.toUpperCase().padEnd(SEVERITY_WIDTH);
-  const status = effectiveStatus(release, issue).padEnd(STATUS_WIDTH);
-  const cell = issueCell(release, issue);
-  const inline = inlineRemediation(release, issue);
-  const tail = inline ? ` → ${inline}` : '';
-  return `${severity} ${status} ${cell}${tail}`;
-}
-
-function inlineRemediation(
-  release: AnalyzeProjectRelease,
-  issue: AnalyzeProjectIssue,
-): string | null {
-  if (issue.type !== 'VULNERABILITY') return null;
-  const partial = chooseInlinePartialFix(issue);
-  if (!partial) return null;
-  const total = release.issues.reduce((n, i) => (i.type === 'VULNERABILITY' ? n + 1 : n), 0);
-  const fixed = total - partial.vulnerabilityIds.length;
-  return `${partial.version} (fixes ${fixed}/${total})`;
-}
-
-function chooseInlinePartialFix(issue: AnalyzeProjectIssue): VersionOption | null {
-  const options = issue.versionOptions;
-  if (!options || options.length === 0) {
-    return null;
-  }
-  const partials = options.filter(
-    (o) => o.fixLevel === 'PARTIAL' && !EXCLUDED_DESCRIPTION_CODES.has(o.descriptionCode),
-  );
-  if (partials.length === 0) {
-    return null;
-  }
-  partials.sort(
-    (a, b) => DESCRIPTION_CODE_ORDER[a.descriptionCode] - DESCRIPTION_CODE_ORDER[b.descriptionCode],
-  );
-  return partials[0];
-}
-
-function packageFixLine(vulnerabilities: AnalyzeProjectIssue[]): string | null {
-  const fixes = packageCompleteFixes(vulnerabilities);
-  if (fixes.length === 0) {
-    return null;
-  }
-  const parts = fixes.map(formatVersionOption);
-  return `${bold('Recommended versions without known vulnerabilities:')} ${parts.join(' | ')}`;
-}
-
-function formatVersionOption(option: VersionOption): string {
-  const label = DESCRIPTION_CODE_LABEL[option.descriptionCode];
-  return label ? `${option.version} (${label})` : option.version;
-}
-
-function packageCompleteFixes(vulnerabilities: AnalyzeProjectIssue[]): VersionOption[] {
-  const perIssueCompleteFixes = vulnerabilities.map(completeFixesByVersion);
-  if (perIssueCompleteFixes.some((m) => m.size === 0)) return [];
-
-  const sharedVersions = intersectKeys(perIssueCompleteFixes);
-  if (sharedVersions.size === 0) return [];
-
-  const representatives: VersionOption[] = [];
-  for (const version of sharedVersions) {
-    representatives.push(bestRepresentative(version, perIssueCompleteFixes));
-  }
-  representatives.sort(
-    (a, b) => DESCRIPTION_CODE_ORDER[a.descriptionCode] - DESCRIPTION_CODE_ORDER[b.descriptionCode],
-  );
-  return representatives.slice(0, MAX_PACKAGE_FIXES);
-}
-
-function completeFixesByVersion(issue: AnalyzeProjectIssue): Map<string, VersionOption> {
-  const out = new Map<string, VersionOption>();
-  const options = issue.versionOptions;
-  if (!options) return out;
-  for (const option of options) {
-    if (option.fixLevel !== 'COMPLETE') continue;
-    if (EXCLUDED_DESCRIPTION_CODES.has(option.descriptionCode)) continue;
-    out.set(option.version, option);
-  }
-  return out;
-}
-
-function intersectKeys(maps: Map<string, VersionOption>[]): Set<string> {
-  if (maps.length === 0) return new Set();
-  let intersection = new Set(maps[0].keys());
-  for (let i = 1; i < maps.length; i++) {
-    const next = new Set<string>();
-    for (const key of intersection) {
-      if (maps[i].has(key)) next.add(key);
-    }
-    intersection = next;
-    if (intersection.size === 0) break;
-  }
-  return intersection;
-}
-
-function bestRepresentative(
-  version: string,
-  perIssueCompleteFixes: Map<string, VersionOption>[],
-): VersionOption {
-  // Every map in `perIssueCompleteFixes` contains `version` by construction
-  // (it came from the intersection of their keys), so collect-then-sort is
-  // safe and lets the compiler narrow the result.
-  const candidates: VersionOption[] = [];
-  for (const map of perIssueCompleteFixes) {
-    const option = map.get(version);
-    if (option) candidates.push(option);
-  }
-  if (candidates.length === 0) {
-    throw new Error(`bestRepresentative invariant violated: ${version} missing from inputs`);
-  }
-  candidates.sort(
-    (a, b) => DESCRIPTION_CODE_ORDER[a.descriptionCode] - DESCRIPTION_CODE_ORDER[b.descriptionCode],
-  );
-  return candidates[0];
 }
 
 function appendErrors(lines: string[], errors: AnalysisErrorResource[]): void {
@@ -343,14 +170,14 @@ function appendSummaryBlock(
   packagesScanned: number,
   totalRisks: number,
 ): void {
-  const counts = countsByTypeAndSeverity(releases);
+  const counts = summaryCountsByTypeAndSeverity(releases);
   lines.push('', summaryHeader(packagesScanned, totalRisks));
   for (const type of ISSUE_TYPES) {
     lines.push(summaryLineForType(type, counts.get(type) ?? new Map()));
   }
 }
 
-function countsByTypeAndSeverity(
+function summaryCountsByTypeAndSeverity(
   releases: AnalyzeProjectRelease[],
 ): Map<ScaIssueType, Map<string, number>> {
   const out = new Map<ScaIssueType, Map<string, number>>();
@@ -372,24 +199,13 @@ function countsByTypeAndSeverity(
 }
 
 function summaryLineForType(type: ScaIssueType, counts: Map<string, number>): string {
-  const cells = SEVERITIES.map((sev) => severityCell(sev, counts.get(sev) ?? 0));
+  const cells = SEVERITIES.map((sev) => summarySeverityCell(sev, counts.get(sev) ?? 0));
   return `  ${type.padEnd(TYPE_LABEL_WIDTH)}  ${cells.join('    ')}`;
 }
 
-function severityCell(label: string, count: number): string {
+function summarySeverityCell(label: string, count: number): string {
   const icon = count === 0 ? green(STATUS_ICONS.done) : red(STATUS_ICONS.failed);
   return `${label} ${icon} ${String(count).padStart(3)}`;
-}
-
-function issueCell(release: AnalyzeProjectRelease, issue: AnalyzeProjectIssue): string {
-  switch (issue.type) {
-    case 'VULNERABILITY':
-      return issue.vulnerabilityId ?? '';
-    case 'PROHIBITED_LICENSE':
-      return issue.spdxLicenseId ?? release.licenseExpression ?? '';
-    case 'MALWARE':
-      return 'Malicious package';
-  }
 }
 
 function transitiveChainLines(
