+++ src/cli/commands/analyze/dependency-risk-helpers/table/format-table-license-group.ts
@@ -1,0 +1,40 @@
+/*
+ * SonarQube CLI
+ * Copyright (C) SonarSource Sàrl
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+import { bold } from '../../../../../ui/colors.js';
+import type { AnalyzeProjectIssue, AnalyzeProjectRelease } from '../sca-scanner.ts';
+import { genericRiskInfo } from './format-table-risk-generic.ts';
+
+export function appendLicenseGroup(
+  lines: string[],
+  release: AnalyzeProjectRelease,
+  licenses: AnalyzeProjectIssue[],
+): void {
+  for (const issue of licenses) lines.push(licenseRow(release, issue));
+  lines.push(bold('Review the license usage'));
+}
+
+function licenseRow(release: AnalyzeProjectRelease, issue: AnalyzeProjectIssue): string {
+  return genericRiskInfo(release, issue, licenseCell(release, issue));
+}
+
+function licenseCell(release: AnalyzeProjectRelease, issue: AnalyzeProjectIssue): string {
+  return issue.spdxLicenseId ?? release.licenseExpression ?? '';
+}
