+++ src/cli/commands/analyze/dependency-risk-helpers/table/format-table-vulnerability-group.ts
@@ -1,0 +1,187 @@
+/*
+ * SonarQube CLI
+ * Copyright (C) SonarSource Sàrl
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+import { bold } from '../../../../../ui/colors.js';
+import type {
+  AnalyzeProjectIssue,
+  AnalyzeProjectRelease,
+  VersionOption,
+  VersionOptionDescriptionCode,
+} from '../sca-scanner.ts';
+import { genericRiskInfo } from './format-table-risk-generic.ts';
+
+const DESCRIPTION_CODE_ORDER: Record<VersionOptionDescriptionCode, number> = {
+  LATEST_STABLE: 0,
+  LATEST_COMPLETE: 1,
+  LATEST_PRERELEASE: 2,
+  LATEST_PARTIAL: 3,
+  NEAREST_COMPLETE: 4,
+  NEAREST_PARTIAL: 5,
+  VERSION_IN_USE: 6,
+  UNKNOWN: 7,
+};
+
+const EXCLUDED_DESCRIPTION_CODES: ReadonlySet<VersionOptionDescriptionCode> = new Set([
+  'VERSION_IN_USE',
+  'UNKNOWN',
+]);
+
+const DESCRIPTION_CODE_LABEL: Record<VersionOptionDescriptionCode, string | null> = {
+  LATEST_STABLE: 'latest stable',
+  LATEST_COMPLETE: 'latest',
+  LATEST_PRERELEASE: 'latest prerelease',
+  LATEST_PARTIAL: 'latest',
+  NEAREST_COMPLETE: 'nearest',
+  NEAREST_PARTIAL: 'nearest',
+  VERSION_IN_USE: null,
+  UNKNOWN: null,
+};
+
+const MAX_PACKAGE_FIXES = 2;
+
+export function appendVulnerabilityGroup(
+  lines: string[],
+  release: AnalyzeProjectRelease,
+  vulnerabilities: AnalyzeProjectIssue[],
+): void {
+  const total = vulnerabilities.length;
+  for (const issue of vulnerabilities) lines.push(vulnerabilityRow(release, issue, total));
+  const fixLine = packageFixLine(vulnerabilities);
+  if (fixLine !== null) lines.push(fixLine);
+}
+
+function vulnerabilityRow(
+  release: AnalyzeProjectRelease,
+  issue: AnalyzeProjectIssue,
+  totalVulnerabilities: number,
+): string {
+  const cell = vulnerabilityCell(issue);
+  const tail = vulnerabilityInlineTail(issue, totalVulnerabilities);
+  return genericRiskInfo(release, issue, tail ? `${cell} → ${tail}` : cell);
+}
+
+function vulnerabilityCell(issue: AnalyzeProjectIssue): string {
+  return issue.vulnerabilityId ?? '';
+}
+
+function vulnerabilityInlineTail(
+  issue: AnalyzeProjectIssue,
+  totalVulnerabilities: number,
+): string | null {
+  const partial = chooseInlinePartialFix(issue);
+  if (!partial) return null;
+  const fixed = totalVulnerabilities - partial.vulnerabilityIds.length;
+  return `${partial.version} (fixes ${fixed}/${totalVulnerabilities})`;
+}
+
+function chooseInlinePartialFix(issue: AnalyzeProjectIssue): VersionOption | null {
+  const options = issue.versionOptions;
+  if (!options || options.length === 0) {
+    return null;
+  }
+  const partials = options.filter(
+    (o) => o.fixLevel === 'PARTIAL' && !EXCLUDED_DESCRIPTION_CODES.has(o.descriptionCode),
+  );
+  if (partials.length === 0) {
+    return null;
+  }
+  partials.sort(
+    (a, b) => DESCRIPTION_CODE_ORDER[a.descriptionCode] - DESCRIPTION_CODE_ORDER[b.descriptionCode],
+  );
+  return partials[0];
+}
+
+function packageFixLine(vulnerabilities: AnalyzeProjectIssue[]): string | null {
+  const fixes = packageCompleteFixes(vulnerabilities);
+  if (fixes.length === 0) {
+    return null;
+  }
+  const parts = fixes.map(formatVersionOption);
+  return `${bold('Recommended versions without known vulnerabilities:')} ${parts.join(' | ')}`;
+}
+
+function formatVersionOption(option: VersionOption): string {
+  const label = DESCRIPTION_CODE_LABEL[option.descriptionCode];
+  return label ? `${option.version} (${label})` : option.version;
+}
+
+function packageCompleteFixes(vulnerabilities: AnalyzeProjectIssue[]): VersionOption[] {
+  const perIssueCompleteFixes = vulnerabilities.map(completeFixesByVersion);
+  if (perIssueCompleteFixes.some((m) => m.size === 0)) return [];
+
+  const sharedVersions = intersectKeys(perIssueCompleteFixes);
+  if (sharedVersions.size === 0) return [];
+
+  const representatives: VersionOption[] = [];
+  for (const version of sharedVersions) {
+    representatives.push(bestRepresentative(version, perIssueCompleteFixes));
+  }
+  representatives.sort(
+    (a, b) => DESCRIPTION_CODE_ORDER[a.descriptionCode] - DESCRIPTION_CODE_ORDER[b.descriptionCode],
+  );
+  return representatives.slice(0, MAX_PACKAGE_FIXES);
+}
+
+function completeFixesByVersion(issue: AnalyzeProjectIssue): Map<string, VersionOption> {
+  const out = new Map<string, VersionOption>();
+  const options = issue.versionOptions;
+  if (!options) return out;
+  for (const option of options) {
+    if (option.fixLevel !== 'COMPLETE') continue;
+    if (EXCLUDED_DESCRIPTION_CODES.has(option.descriptionCode)) continue;
+    out.set(option.version, option);
+  }
+  return out;
+}
+
+function intersectKeys(maps: Map<string, VersionOption>[]): Set<string> {
+  if (maps.length === 0) return new Set();
+  let intersection = new Set(maps[0].keys());
+  for (let i = 1; i < maps.length; i++) {
+    const next = new Set<string>();
+    for (const key of intersection) {
+      if (maps[i].has(key)) next.add(key);
+    }
+    intersection = next;
+    if (intersection.size === 0) break;
+  }
+  return intersection;
+}
+
+function bestRepresentative(
+  version: string,
+  perIssueCompleteFixes: Map<string, VersionOption>[],
+): VersionOption {
+  // Every map in `perIssueCompleteFixes` contains `version` by construction
+  // (it came from the intersection of their keys), so collect-then-sort is
+  // safe and lets the compiler narrow the result.
+  const candidates: VersionOption[] = [];
+  for (const map of perIssueCompleteFixes) {
+    const option = map.get(version);
+    if (option) candidates.push(option);
+  }
+  if (candidates.length === 0) {
+    throw new Error(`bestRepresentative invariant violated: ${version} missing from inputs`);
+  }
+  candidates.sort(
+    (a, b) => DESCRIPTION_CODE_ORDER[a.descriptionCode] - DESCRIPTION_CODE_ORDER[b.descriptionCode],
+  );
+  return candidates[0];
+}
